move navItems to lib
