+++ components/Connect.jsx
@@ -1,10 +1,10 @@
 import A from '@components/A'
+import { navItems } from '@lib/navItems'
+
 export default () => (
-    <div className="text-gray-600 text-xl flex flex-col space-y-2">
-      <A href='mailto:davidjamesmagnuson@gmail.com' text='email'/>
-      <A href='https://github.com/davidbrekke' text='github'/>
-      <A href='https://mirror.xyz/0xBc08269F39757bd007815288D2717fFd77F75B77' text='mirror'/>
-      <A href='https://twitter.com/d_mgnsn' text='twitter'/>
-      <A href='https://www.instagram.com/david_magnuson/' text='instagram'/>
-    </div>
+  <div className="text-gray-600 text-xl flex flex-col space-y-2">
+    {navItems.map(({ url, title }, i) => (
+      <A href={url} text={title} key={i} />
+    ))}
+  </div>
 )
