+++ components/layout/Nav.jsx
@@ -1,20 +1,10 @@
 import { useState } from 'react'
 import Link from 'next/link'
 import { MenuAlt3Icon } from '@heroicons/react/solid'
+import { navItems } from '@lib/navItems'
 
 const Nav = () => {
   const [isOpen, setIsOpen] = useState(false)
-
-  const navItems = [
-    { url: 'mailto:davidjamesmagnuson@gmail.com', title: 'email' },
-    { url: 'https://github.com/davidbrekke', title: 'github' },
-    { url: 'https://mirror.xyz/0xBc08269F39757bd007815288D2717fFd77F75B77', title: 'mirror' },
-    { url: 'https://twitter.com/d_mgnsn', title: 'twitter' },
-    { url: 'https://www.instagram.com/david_magnuson/', title: 'instagram' },
-    // { url: '/react', title: 'react' },
-    // { url: '/blog', title: 'blog' },
-    // { url: '/contact', title: 'contact' },
-  ]
 
   return (
     <>
@@ -30,7 +20,10 @@ const Nav = () => {
       >
         {navItems.map(({ title, url }, i) => (
           <Link href={url} key={i}>
-            <a target="_blank" className="hover:text-indigo-400 cursor-pointer transition">
+            <a
+              target="_blank"
+              className="hover:text-indigo-400 cursor-pointer transition"
+            >
               {title}
             </a>
           </Link>
