+++ lib/navItems.js
@@ -1,0 +1,12 @@
+const navItems = [
+  { url: 'mailto:davidjamesmagnuson@gmail.com', title: 'email' },
+  { url: 'https://github.com/davidbrekke', title: 'github' },
+  {
+    url: 'https://mirror.xyz/0xBc08269F39757bd007815288D2717fFd77F75B77',
+    title: 'mirror',
+  },
+  { url: 'https://twitter.com/d_mgnsn', title: 'twitter' },
+  { url: 'https://www.instagram.com/david_magnuson/', title: 'instagram' },
+]
+
+export { navItems }
